refactor: index.js 활용하여 검색 용이하게 하기
