+++ src/components/Home/Article/index.js
@@ -1,0 +1,3 @@
+import Article from "./Article";
+
+export default Article;
