+++ src/components/Home/HomeStories.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import styled from "styled-components";
-import Card from "UI/Card/Card";
+import Card from "UI/Card";
 import Story from "./Story";
 import sprite from "../../assets/Images/sprite.png";
