+++ src/components/Home/Modals/ReportModal/index.js
@@ -1,0 +1,3 @@
+import ReportModal from "./ReportModal";
+
+export default ReportModal;
